# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Amazing Studio - Wedding Studio Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    put:
      operationId: updateStaff
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [package, addon]
        price:
          type: number
        duration:
          type: string
          nullable: true
        includes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, price, includes, isActive, createdAt]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [package, addon]
        price:
          type: number
        duration:
          type: string
          nullable: true
        includes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [name, type, price, includes, isActive]
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, photographer, assistant, receptionist]
        email:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        joinDate:
          type: string
          format: date
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, role, joinDate, isActive, createdAt]
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, photographer, assistant, receptionist]
        email:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        joinDate:
          type: string
          format: date
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
      required: [name, phone, role, joinDate, isActive]
...